fix bug

+++ hw4/hw4.py
@@ -43,11 +43,12 @@ def preprocess_data():
     def make_subset(subset_name, start_index, end_index):
         for category in ("cat", "dog"):
             dir = new_base_dir / subset_name / category
-            os.makedirs(dir)
-            fnames = [f"{category}.{i}.jpg" for i in range(start_index, end_index)]
-            for fname in fnames:
-                shutil.copyfile(src=original_dir / fname,
-                                dst=dir / fname)
+            if(not os.path.exists(dir)):
+                os.makedirs(dir)
+                fnames = [f"{category}.{i}.jpg" for i in range(start_index, end_index)]
+                for fname in fnames:
+                    shutil.copyfile(src=original_dir / fname,
+                                    dst=dir / fname)
     make_subset("train", start_index=0, end_index=1000)
     make_subset("validation", start_index=1000, end_index=1500)
     make_subset("test", start_index=1500, end_index=2500)
